+++ src/main/java/net/pwall/log/ConsoleLoggerFactory.java
@@ -2,7 +2,7 @@
  * @(#) ConsoleLoggerFactory.java
  *
  * log-front  Logging interface
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -28,7 +28,7 @@ package net.pwall.log;
 import java.io.PrintStream;
 
 /**
- * A LoggerFactory that creates {@link ConsoleLogger} objects.
+ * A {@link LoggerFactory} that creates {@link ConsoleLogger} objects.
  *
  * @author  Peter Wall
  */