+++ src/main/java/net/pwall/log/NullLoggerFactory.java
@@ -2,7 +2,7 @@
  * @(#) NullLoggerFactory.java
  *
  * log-front  Logging interface
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,7 +26,7 @@
 package net.pwall.log;
 
 /**
- * A LoggerFactory that returns only a {@link NullLogger}.
+ * A {@link LoggerFactory} that returns only a {@link NullLogger}.
  *
  * @author  Peter Wall
  */