+++ src/main/java/net/pwall/log/Slf4JLoggerException.java
@@ -2,7 +2,7 @@
  * @(#) Slf4JLoggerException.java
  *
  * log-front  Logging interface
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,7 +26,7 @@
 package net.pwall.log;
 
 /**
- * An exception class to handle errors in accessing slf4j.
+ * An exception class to handle errors in accessing {@code slf4j}.
  *
  * @author  Peter Wall
  */